backend/requirements.txt
# FastAPI & Web
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
websockets>=12.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.0

# Data & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
ta>=0.10.0

# Notifications
python-telegram-bot>=20.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-dotenv>=1.0.0

# Redis
redis>=5.0.0

# Utils
pydantic>=2.0.0
aiohttp>=3.8.0
matplotlib>=3.7.0
